quote-service/requirements.txt
amqp==5.1.1
aniso8601==9.0.1
attrs==21.4.0
beautifulsoup4==4.11.1
billiard==3.6.4.0
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cycler==0.11.0
Flask==2.1.2
Flask-Cors==3.0.10
flask-restx==0.5.1
fonttools==4.33.3
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.6.0
kiwisolver==1.4.3
kombu==5.2.4
MarkupSafe==2.1.1
matplotlib==3.5.2
numpy
packaging==21.3
pandas
Pillow==9.1.1
prompt-toolkit==3.0.29
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
six==1.16.0
soupsieve==2.3.2.post1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.2
zipp==3.8.0
